Removal of Px usages part 6 of n
Relnote: "Replaced usage of Px class in various
compose classes as part of the large
refactoring effort to only rely on Dp
and primitive types for pixel parameters"

Bug: 156681014
Test: re-ran compose tests

+++ ui/ui-material/src/androidTest/java/androidx/ui/material/AppBarTest.kt
@@ -144,7 +144,7 @@ class AppBarTest {
             val actionPositionY = actionCoords!!.globalPosition.y
             val actionExpectedPositionY = (appBarBottomEdgeY - AppBarStartAndEndPadding.toPx() -
                 FakeIconSize.toPx()
-            ).roundToInt()
+            ).roundToInt().toFloat()
             assertThat(actionPositionY).isEqualTo(actionExpectedPositionY)
         }
     }
